Create Convert_Election_Committee_Expenses
Here is a method to convert each election committee’s “Operating Expenditures” in 2016 to CSV format.  Election committees are required to report their operating expenses to the Federal Election Commission on FEC Form 3.   Each item includes a description of what the expense was for, the date it was paid, the amount paid, and to whom it was paid.  This will allow you to look at every itemized individual expense by each election committee in 2016 in CSV format. 
#This methodology should also work with all FEC files dating back to the year 2000.